+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 Versions follow [Semantic Versioning](https://semver.org/spec/v2.0.0.html) (`<major>`.`<minor>`.`<patch>`)
 
+## [v2.2.0]
+### Added
+* #87 Add `--allow-untyped-defs` to suppress all errors from dynamically typed functions. A function is considered dynamically typed if it does not contain any type hints.
+
+### Fixed
+* #77, #81 Fix incorrect return error locations in the presence of a multiline docstring containing colon(s)
+* #81 Fix incorrect return error locations for single-line function definitions
+* Fix incorrectly pinned project specifications
+
 ## [v2.1.0]
 ### Added
 * #68 Add `--suppress-dummy-args` configuration option to suppress ANN000 level errors for dummy arguments, defined as `"_"`